requirements.txt
pycryptodomex
rich